Mooved the Search bar to the top - Need to fix CSS
