+++ src/components/Siders/SearchSider.css
@@ -1,5 +1,5 @@
 #sider {
-  background-image: url("https://res.cloudinary.com/dd5irg4pr/image/upload/v1669128876/background_jobpag.png");
+  /* background-image: url("https://res.cloudinary.com/dd5irg4pr/image/upload/v1669128876/background_jobpag.png"); */
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
