+++ src/components/Siders/SimpleSider.css
@@ -1,5 +1,5 @@
 #simpleSider {
-    background-image: url('https://res.cloudinary.com/dd5irg4pr/image/upload/v1669128876/background_jobpag.png');
+    /* background-image: url('https://res.cloudinary.com/dd5irg4pr/image/upload/v1669128876/background_jobpag.png'); */
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
